pyproject.toml
[project]
name = "hypercube-python-operators"
version = "0.1.0"
description = "Basic python operators for the hypercube project"
authors = [{ name = "Alejandro Agustin" }]
requires-python = "~=3.12"
readme = "README.md"
dependencies = [
    "pandas>=2.2.3,<3",
    "pydantic>=2.10.4,<3",
    "pyarrow>=18.1.0,<19",
    "pyyaml==6.0.2",
]

[dependency-groups]
dev = [
    "coverage>=7.5.2,<8",
    "pytest>=8.2.1,<9",
    "ruff>=0.4.5,<0.5",
    "bandit>=1.7.8,<2",
    "black>=24,<25",
    "setuptools>=75.8.0",
]

[tool.hatch.build.targets.wheel]
      packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
